Documentation and testing

+++ app/Http/ApiV1/Modules/Games/Controllers/GamesController.php
@@ -6,6 +6,7 @@ use App\Domain\Create\Actions\AuthorizeAction;
 use App\Domain\Create\Actions\ConnectGameAction;
 use App\Domain\Create\Actions\CreateGameAction;
 use App\Domain\Create\Models\Game;
+use App\Http\ApiV1\Modules\Games\Requests\GameConnectOrDestroyRequest;
 use App\Http\ApiV1\Modules\Games\Requests\GameRequest;
 use App\Http\ApiV1\Modules\Games\Resources\GameResource;
 use Exception;
@@ -17,6 +18,7 @@ use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 use Illuminate\Validation\UnauthorizedException;
 use Symfony\Component\HttpFoundation\Response as ResponseAlias;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class GamesController extends Controller
 {
@@ -28,7 +30,7 @@ class GamesController extends Controller
         try {
             return GameResource::collection(Game::all());
         } catch (Exception $e) {
-            return response()->json($e->getMessage(), 404);
+            return response()->json(['errors' => [['code' => "401", 'message' => $e->getMessage()]]], 404);
         }
     }
 
@@ -42,7 +44,7 @@ class GamesController extends Controller
             $authorized = $authorizeAction->execute($validate);
             $game = $gameAction->execute($authorized);
         } catch (Exception $e) {
-            return response()->json(['error' => ['message' => $e->getMessage()]], 401);
+            return response()->json(['data' => [], 'errors' => [['code' => "401", 'message' => $e->getMessage()]]], 401);
         }
 
         return new GameResource($game);
@@ -52,7 +54,7 @@ class GamesController extends Controller
      * Connect to the game.
      */
     public function connect(
-        GameRequest $request,
+        GameConnectOrDestroyRequest $request,
         ConnectGameAction $connectAction,
         AuthorizeAction $authorizeAction
     ): JsonResponse|\Illuminate\Contracts\Foundation\Application|ResponseFactory|Application|Response {
@@ -61,7 +63,7 @@ class GamesController extends Controller
             $authorized = $authorizeAction->execute($validate);
             $connectAction->execute($authorized);
         } catch (Exception $e) {
-            return response()->json(['error' => ['message' => $e->getMessage()]], 404);
+            return response()->json(['data' => [], 'errors' => [['code' => $e->getCode(), 'message' => $e->getMessage()]]], $e->getCode());
         }
 
         return response(null, ResponseAlias::HTTP_ACCEPTED);
@@ -70,19 +72,22 @@ class GamesController extends Controller
     /**
      * Delete the game.
      */
-    public function destroy(GameRequest $request, AuthorizeAction $authorizeAction): JsonResponse|\Illuminate\Contracts\Foundation\Application|ResponseFactory|Application|Response
+    public function destroy(GameConnectOrDestroyRequest $request, AuthorizeAction $authorizeAction): JsonResponse|\Illuminate\Contracts\Foundation\Application|ResponseFactory|Application|Response
     {
         try {
             $validate = $request->validated();
             $authorized = $authorizeAction->execute($validate);
             $game = Game::find($authorized['id']);
+            if ($game === null) {
+                throw new NotFoundResourceException("Game not found", 404);
+            }
             if ($authorized['user_id'] == $game->player_white || $authorized['user_id'] == $game->player_black) {
                 $game->delete();
             } else {
-                throw new UnauthorizedException("Not a game owner");
+                throw new UnauthorizedException("Not a game owner", 401);
             }
         } catch (Exception $e) {
-            return response()->json(['error' => ['message' => $e->getMessage()]], 400);
+            return response()->json(['data' => [], 'errors' => [['code' => $e->getCode(), 'message' => $e->getMessage()]]], $e->getCode());
         }
 
         return response(null, ResponseAlias::HTTP_NO_CONTENT);
